libraries combined
