--- rails/crowdsource-due-diligence/lib/positive_library.rb
@@ -1,7 +1,0 @@
-module PositiveLibrary
-
-  def positive_library
-    file='lib/opinion_lexicon_english/positive_words.txt'
-    File.readlines(file).collect{|line| line.chomp}
-  end
-end
